refactor: extract shared wrapper envelope to src/wrapper.ts
Pulls cassette lookup, mode resolution, ack gate, matcher init,
and record/replay branching into a runner-agnostic envelope.
src/execa.ts becomes a thin entry that wires execa-specific bits
via the internal RunnerHooks type. No behavior change. All v0.1
tests remain green.

Sets up for v0.2's tinyexec adapter to share the envelope.
